Refactor Config Window + version bumps
# Config Window
  1. Created 5 single-responsibility submodules:
      - dialog.rs: DialogState enum & action dialog modal overlays.
      - search_bar.rs: Search bar widget & fuzzy match filter logic.
      - item_row.rs: Item row container styling & folder button style helpers.
      - playlist_view.rs: Playlist tab, empty state, and song row views.
      - statics_view.rs: Statics tab and static row views.
  2. Refactored mod.rs:
      - Reduced from 701 lines to 109 lines.
      - Preserved all public API re-exports (DialogState, PLAYLIST_SCROLLABLE_ID).
      - Replaced inner UI construction methods with delegation to the extracted submodules.

# Version Bumps
- id3 to 1.17
- tokio to 1.53
- uuid to 1.24
