+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2024"
 authors = ["Klassenserver7b <klassenserver7bwin10@gmail.com>", "Konsl <82901383+Konsl@users.noreply.github.com>, Robocraft999 <59525378+Robocraft999@users.noreply.github.com>"]
 
 [dependencies]
-id3 = "1.16"
+id3 = "1.17"
 url = "2.5"
 percent-encoding = "2.3"
 iced = { version = "0.14.0", features = ["linux-theme-detection", "image", "wayland", "svg", "lazy", "canvas", "tokio", "advanced"] }
@@ -16,12 +16,12 @@ iced_aw = { version = "0.14.1", features = ["default", "color_picker"] }
 rfd = "0.17"
 dirs = "6.0"
 warp = { version = "0.4", features = ["websocket", "server"] }
-tokio = { version = "1.51" }
+tokio = { version = "1.53" }
 pin-project-lite = "0.2"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 indexmap = { version = "2.7", features = ["serde"] }
-uuid = { version = "1.23", features = ["v4"] }
+uuid = { version = "1.24", features = ["v4"] }
 bytes = "1.11"
 network-interface = "2.0"
 fuzzy-matcher = "0.3"
